refactor(settings): extract multi-host config model
